+++ Changelog.md
@@ -15,6 +15,7 @@
 - [#609]: Added `Writer::write_serializable` to provide the capability to serialize
   arbitrary types using serde when using the lower-level `Writer` API.
 - [#615]: Added ability to set entity resolver when deserialize using borrowing reader.
+- [#617]: Added ability to enforce the expansion of empty elements.
 
 ### Bug Fixes
 
@@ -26,6 +27,7 @@
 [#604]: https://github.com/tafia/quick-xml/issue/604
 [#609]: https://github.com/tafia/quick-xml/pull/609
 [#615]: https://github.com/tafia/quick-xml/pull/615
+[#617]: https://github.com/tafia/quick-xml/pull/617
 
 
 ## 0.29.0 -- 2023-06-13